monorepo/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
ollama==0.3.2
python-multipart==0.0.9
jinja2==3.1.4
pypdf==5.0.1
python-docx==1.1.2
beautifulsoup4==4.12.3
lxml==5.3.0
chardet==5.2.0
